esg-sustainability-data-platform-main/root_app/requirements.txt
matplotlib
plotly
hydralit-components
snowflake-connector-python
snowflake-snowpark-python
streamlit
streamlit-authenticator
streamlit_scrollable_textbox